refactor: group repositories by bounded context 🚤
