+++ CoffeeChess.Infrastructure/Repositories/Implementations/Matchmaking/RedisChallengeRepository.cs
@@ -1,13 +1,11 @@
-﻿using System.Text.Json;
-using CoffeeChess.Domain.Matchmaking.Entities;
+﻿using CoffeeChess.Domain.Matchmaking.Entities;
 using CoffeeChess.Domain.Matchmaking.Enums;
 using CoffeeChess.Domain.Matchmaking.Repositories.Interfaces;
 using CoffeeChess.Domain.Matchmaking.ValueObjects;
 using CoffeeChess.Infrastructure.Exceptions;
-using CoffeeChess.Infrastructure.Serialization;
 using StackExchange.Redis;
 
-namespace CoffeeChess.Infrastructure.Repositories.Implementations;
+namespace CoffeeChess.Infrastructure.Repositories.Implementations.Matchmaking;
 
 public class RedisChallengeRepository(
     IConnectionMultiplexer redis) : IChallengeRepository
