+++ CoffeeChess.Web/Program.cs
@@ -18,6 +18,11 @@ using CoffeeChess.Domain.Players.Services.Interfaces;
 using CoffeeChess.Infrastructure.Identity;
 using CoffeeChess.Infrastructure.Persistence;
 using CoffeeChess.Infrastructure.Repositories.Implementations;
+using CoffeeChess.Infrastructure.Repositories.Implementations.Chats;
+using CoffeeChess.Infrastructure.Repositories.Implementations.Games;
+using CoffeeChess.Infrastructure.Repositories.Implementations.Matchmaking;
+using CoffeeChess.Infrastructure.Repositories.Implementations.Players;
+using CoffeeChess.Infrastructure.Repositories.Implementations.Songs;
 using CoffeeChess.Infrastructure.Serialization;
 using CoffeeChess.Infrastructure.Services.Implementations;
 using CoffeeChess.Web.BackgroundWorkers;
